Rename Class -> ClassNode and extract FunctionNode from Function*s again
