+++ src/ast/nodes/ClassDeclaration.js
@@ -1,6 +1,6 @@
-import Class from './shared/Class.js';
+import ClassNode from './shared/ClassNode.js';
 
-export default class ClassDeclaration extends Class {
+export default class ClassDeclaration extends ClassNode {
 	initialiseChildren ( parentScope ) {
 		// Class declarations are like let declarations: Not hoisted, can be reassigned, cannot be redeclared
 		this.id && this.id.initialiseAndDeclare( parentScope, 'class', this );
