+++ src/ast/nodes/ClassExpression.js
@@ -1,6 +1,6 @@
-import Class from './shared/Class.js';
+import ClassNode from './shared/ClassNode.js';
 
-export default class ClassExpression extends Class {
+export default class ClassExpression extends ClassNode {
 	initialiseChildren ( parentScope ) {
 		this.id && this.id.initialiseAndDeclare( this.scope, 'class', this );
 		super.initialiseChildren( parentScope );
